Add manual checkout method
Streaming responses don't work with the Crystal type system. The `yield`
inside of the `exec` overload that receives a block that is passed to an
`HTTP::Client` inside. Since `HTTPClient::Client` is itself an
`HTTP::Client`, the Crystal compiler sees that it could yield to an
instance of the same class, infinitely, and avoids trying to compile it.